+++ packages/server/src/routes/sessionRoutes.ts
@@ -74,6 +74,10 @@ export const getSessionRoute = createRoute({
       content: { 'application/json': { schema: GetSessionResponseSchema } },
       description: 'Session data.',
     },
+    403: {
+      content: { 'application/json': { schema: RequestErrorResponseSchema } },
+      description: 'Caller is not the session creator.',
+    },
     404: {
       content: { 'application/json': { schema: RequestErrorResponseSchema } },
       description: 'Session not found.',
@@ -142,7 +146,7 @@ export const listSessionsRoute = createRoute({
   tags: [SESSIONS_TAG],
   summary: 'List sessions',
   description:
-    'List sessions (newest first by default), token-paginated. Optional `agent_id` filters to sessions bound to that named agent; optional `created_by` filters by creator identity. Pass `page_token` to fetch the next page, keeping the other query params constant.',
+    "List the caller's sessions (newest first by default), token-paginated. Results are scoped to the authenticated identity via the session store's `created_by` filter (not a client query param). Optional `agent_id` filters to sessions bound to that named agent. Pass `page_token` to fetch the next page, keeping the other query params constant.",
   'x-fern-sdk-group-name': ['sessions'],
   'x-fern-sdk-method-name': 'list',
   'x-fern-pagination': TOKEN_PAGINATION,
