+++ packages/server/tests/unit/apis/deletedSessionCrud.test.ts
@@ -5,6 +5,7 @@ import { createClient } from 'redis';
 import { createLogger } from 'winston';
 import { createSessionsRouter, TENANT_ID } from '../../../src/apis/sessions';
 import { createTurnsRouter } from '../../../src/apis/turns';
+import { LOCAL_USER_CONTEXT } from '../../../src/auth/identity';
 import { migrateSqliteToLatest } from '../../../src/db/migrateSqlite';
 import { SqliteAgentStore } from '../../../src/db/sqlite/agent-store/SqliteAgentStore';
 import { createSqliteDb } from '../../../src/db/sqlite/client';
@@ -68,7 +69,7 @@ describe('public CRUD after session deletion', () => {
     await sessionStore.createSession({
       tenant_id: TENANT_ID,
       session_id: 's1',
-      created_by: 'user-1',
+      created_by: LOCAL_USER_CONTEXT.userRef,
       agent: {
         type: 'value',
         agent_spec: AgentSpecSchema.parse({
