+++ packages/server/tests/unit/apis/sessionHttp.test.ts
@@ -3,6 +3,7 @@ import { AgentSpecSchema, Sessions } from '@truefoundry/utils-core/agent-session
 import { RequestReplyRouter } from '@truefoundry/utils-core/request-reply';
 import { createClient } from 'redis';
 import { createSessionsRouter, TENANT_ID } from '../../../src/apis/sessions';
+import { LOCAL_USER_CONTEXT } from '../../../src/auth/identity';
 import { migrateSqliteToLatest } from '../../../src/db/migrateSqlite';
 import { SqliteAgentStore } from '../../../src/db/sqlite/agent-store/SqliteAgentStore';
 import { createSqliteDb } from '../../../src/db/sqlite/client';
@@ -29,11 +30,12 @@ function jsonInit(method: string, body: unknown): RequestInit {
 describe('sessions HTTP agent binding', () => {
   let app: OpenAPIHono;
   let agentStore: SqliteAgentStore;
+  let sessionStore: SqliteSessionStore;
 
   beforeEach(async () => {
     const db = createSqliteDb(':memory:');
     await migrateSqliteToLatest(db);
-    const sessionStore = new SqliteSessionStore(db);
+    sessionStore = new SqliteSessionStore(db);
     const sessions = new Sessions({ sessionStore });
     const modelProviderStore = new SqliteModelProviderStore(db);
     const mcpServerStore = new SqliteMcpServerStore(db);
@@ -88,7 +90,7 @@ describe('sessions HTTP agent binding', () => {
     };
     expect(json.data.agent.type).toBe('value');
     expect(json.data.agent.agent_spec?.instructions).toBe('draft');
-    expect(json.data.created_by).toBe('trueforge-default');
+    expect(json.data.created_by).toBe(LOCAL_USER_CONTEXT.userRef);
   });
 
   it('returns 404 when creating a session for an unknown agent_id', async () => {
@@ -122,22 +124,34 @@ describe('sessions HTTP agent binding', () => {
     expect(listJson.data.some(row => row.id === json.data.id)).toBe(true);
   });
 
-  it('filters list by created_by', async () => {
+  it("lists only the caller's sessions and rejects get of another user's session", async () => {
+    await sessionStore.createSession({
+      tenant_id: TENANT_ID,
+      session_id: 'other-user-session',
+      created_by: 'someone-else',
+      agent: { type: 'value', agent_spec: draftSpec },
+      custom: null,
+    });
+
     const created = await app.request('/', jsonInit('POST', { agent: { type: 'value', agent_spec: draftSpec } }));
     expect(created.status).toBe(201);
     const json = (await created.json()) as { data: { id: string; created_by: string } };
-    expect(json.data.created_by).toBe('trueforge-default');
+    expect(json.data.created_by).toBe(LOCAL_USER_CONTEXT.userRef);
 
-    const matched = await app.request('/?created_by=trueforge-default');
-    expect(matched.status).toBe(200);
-    const matchedJson = (await matched.json()) as { data: Array<{ id: string; created_by: string }> };
-    expect(matchedJson.data.some(row => row.id === json.data.id)).toBe(true);
-    expect(matchedJson.data.every(row => row.created_by === 'trueforge-default')).toBe(true);
+    const listed = await app.request('/');
+    expect(listed.status).toBe(200);
+    const listedJson = (await listed.json()) as { data: Array<{ id: string; created_by: string }> };
+    expect(listedJson.data.map(row => row.id)).toEqual([json.data.id]);
+    expect(listedJson.data.every(row => row.created_by === LOCAL_USER_CONTEXT.userRef)).toBe(true);
 
-    const unmatched = await app.request('/?created_by=someone-else');
-    expect(unmatched.status).toBe(200);
-    const unmatchedJson = (await unmatched.json()) as { data: Array<{ id: string }> };
-    expect(unmatchedJson.data.some(row => row.id === json.data.id)).toBe(false);
+    const forbidden = await app.request('/other-user-session');
+    expect(forbidden.status).toBe(403);
+    expect(await forbidden.json()).toEqual({
+      error: { message: 'Only the session creator can get this session' },
+    });
+
+    const allowed = await app.request(`/${json.data.id}`);
+    expect(allowed.status).toBe(200);
   });
 
   it('rejects PATCH agent on a session bound by agent_id', async () => {
